examples/requirements.txt
transformers
huggingface-hub

torch
sentencepiece
pandas
requests
colorama
python-dotenv